@@ -11,30 +11,33 @@
   },
   "dependencies": {
     "@nextui-org/react": "^2.4.8",
-    "@reduxjs/toolkit": "^2.2.7",
-    "@types/react-redux": "^7.1.33",
+    "@reduxjs/toolkit": "^2.3.0",
+    "@types/react-redux": "^7.1.34",
+    "autoprefixer": "^10.4.20",
     "framer-motion": "^11.11.17",
     "localforage": "^1.10.0",
     "match-sorter": "^6.3.4",
     "motion": "^11.11.17",
+    "next-themes": "^0.4.3",
+    "postcss": "^8.4.49",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-redux": "^9.1.2",
-    "react-router-dom": "^6.25.1",
+    "react-router-dom": "^6.28.0",
     "sort-by": "^1.2.0"
   },
   "devDependencies": {
-    "@types/node": "^22.0.0",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
-    "@typescript-eslint/eslint-plugin": "^7.15.0",
-    "@typescript-eslint/parser": "^7.15.0",
-    "@vitejs/plugin-react-swc": "^3.5.0",
-    "eslint": "^8.57.0",
+    "@types/node": "^22.9.1",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
+    "@typescript-eslint/eslint-plugin": "^7.18.0",
+    "@typescript-eslint/parser": "^7.18.0",
+    "@vitejs/plugin-react-swc": "^3.7.1",
+    "eslint": "^8.57.1",
     "eslint-plugin-react-hooks": "^4.6.2",
-    "eslint-plugin-react-refresh": "^0.4.7",
+    "eslint-plugin-react-refresh": "^0.4.14",
     "tailwindcss": "^3.4.15",
-    "typescript": "^5.2.2",
-    "vite": "^5.3.4"
+    "typescript": "^5.6.3",
+    "vite": "^5.4.11"
   }
 }
