@@ -1,16 +1,17 @@
-// tailwind.config.js
+/** @type {import('tailwindcss').Config} */
+
 const { nextui } = require("@nextui-org/react");
 
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-    content: [
-        // ...
-        // make sure it's pointing to the ROOT node_module
-        "./node_modules/@nextui-org/theme/dist/**/*.{js,ts,jsx,tsx}",
-    ],
-    theme: {
-        extend: {},
-    },
-    darkMode: "class",
-    plugins: [nextui()],
-};
+export default {
+  content: [
+    "./index.html",
+    "./src/**/*.{js,ts,jsx,tsx}",
+    "./node_modules/@nextui-org/theme/dist/**/*.{js,ts,jsx,tsx}"
+  ],
+  theme: {
+    extend: {},
+  },
+  darkMode: "class",
+  plugins: [nextui()],
+}
+
