@@ -3,15 +3,13 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App.tsx'
 import './index.css'
-import { Provider } from 'react-redux'
-import store from '@/stores'
 const rootContainer = ReactDOM.createRoot(document.getElementById('root')!);
 rootContainer.render(
   <React.StrictMode>
     <NextUIProvider>
-      <main className="light text-foreground bg-background">
+      <div className="dark text-foreground bg-background h-dvh w-full">
         <App />
-      </main>
+      </div>
     </NextUIProvider>
   </React.StrictMode>,
 )
